kboot: Start to move efi common routines to libkboot
Start to move the common efi routines into libkboot by moving the efi
memory map walking and implementing a printing routine around it.

Sponsored by:		Netflix
