+++ stand/kboot/include/efi.h
@@ -1,0 +1,17 @@
+/*-
+ * Copyright (c) 2024, Netflix, Inc.
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <sys/efi.h>
+#include <machine/metadata.h>
+
+/* Note, we mix and match FreeBSD types and EFI standard defined types */
+
+typedef void (*efi_map_entry_cb)(struct efi_md *, void *argp);
+
+void foreach_efi_map_entry(struct efi_map_header *efihdr, efi_map_entry_cb cb, void *argp);
+void print_efi_map(struct efi_map_header *efihdr);
