+++ stand/kboot/libkboot/Makefile
@@ -13,6 +13,10 @@ SRCS+=	seg.c
 SRCS+=	termios.c
 SRCS+=	util.c
 
+.if ${MACHINE_ARCH} != "powerpc64"
+SRCS+=	efi.c
+.endif
+
 .sinclude "${.CURDIR}/arch/${MACHINE_ARCH}/Makefile.inc"
 
 .include <bsd.lib.mk>
